backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
google-generativeai==0.3.2
pymongo==4.6.1
pypdf2==3.0.1
python-dotenv==1.0.0
sentence-transformers==2.3.1
pydantic==2.5.3
pydantic-settings==2.1.0
numpy==1.26.3
aiofiles==23.2.1
